+++ lib/src/models/trace.dart
@@ -10,11 +10,26 @@ class Trace {
   final String name;
   final Map<String, dynamic> attributes = {};
 
+  /// Sets attribute of execution trace.
+  /// [String] id of the trace.
+  /// [String] key of attribute.
+  /// [String] value of attribute.
+  ///
+  /// Please migrate to the App Flows APIs: [APM.startFlow], [APM.setFlowAttribute], and [APM.endFlow].
+  @Deprecated(
+    'Please migrate to the App Flows APIs: APM.startAppFlow, APM.endFlow, and APM.setFlowAttribute. This feature was deprecated in v13.0.0',
+  )
   void setAttribute(String key, String value) {
     APM.setExecutionTraceAttribute(id, key, value);
     attributes[key] = value;
   }
 
+  /// Ends Execution Trace
+  ///
+  /// Please migrate to the App Flows APIs: [APM.startFlow], [APM.setFlowAttribute], and [APM.endFlow].
+  @Deprecated(
+    'Please migrate to the App Flows APIs: APM.startAppFlow, APM.endFlow, and APM.setFlowAttribute. This feature was deprecated in v13.0.0',
+  )
   void end() {
     APM.endExecutionTrace(id);
   }